Main components underway and a basic sim can run.
However, I'm not sending any traffic during the timed portion of the sim - this commit is about having the pieces in place.

The next steps are as follows:
- update initialization to account for bus and flit widths
- create in/out buffers and credit management
- start exchanging packets

As those are added, we'll need to get some additional depth in our routing approach (especially from endpt to endpt) and develop a real arbitration model for within the router.