Starting in on the arbitration API and implementation.
Spending time reviewing how it's done in merlin and how much we want to copy that model (probably semi-closely); have set up a data structure for holding the heads (much like vc_heads in merlin components) of the virtual channels for each port. Next up is checking that the flit can move to its desired destination and creating a data struct so the router has this knowledge and can perform the move.

For now, the packet just sits and does nothing once it is in the routers inVcHeads structure.