Requests to the endpoint NIC are now broken up into flits.
The flits then traverse the NOC one-by-one and we should only be arbitrating the packet once with the HEAD flit.

Also debugged a couple of issues found when trying to send packets to the same src-dest pair back-to-back.

Need to do a clean up and commenting pass.  May find a couple of places where the logic should be improved as well.