Update readme and add a packet id tracker
Cleaned up some of the printouts from the internals using the packet id value (previously had a flit id) - should make it a bit easier to find where a flit is in the log.

Had the realization (and not sure why it took so long) that concentration == local ports so cleaned up some comments.