WIP: Start trying to use merlin test_nic
Requires building out the untimed send/recv code through all of the components (which had not been done initially).  Started to do that here and the code compiles and starts, but getting a vector failure once the endpoint initialization phase starts.  Unfortunately, not sure which vector is failing (i suspect something is out of range and thus causing issues).