Started testing for multiple VNs.
Basic test works, but need to be able to have an endpoint send/receive cleanly from multiple VNs - tried to hack it a bit in mordredNIC, to no avail.  Was able to count the number of packets received (total across all endpoint) and that matched the expected value, so seems that everything moved around as expected.  I don't think merlin.test_nic supports multiple VNs, so going to have to poke around some.