Add a 3D torus topology.
Need to figure out why the 3D torus is slightly outperforming the 2D torus.  My suspicion is that I have something in the router doing a single port per clock tick and that's changing the timing just enough (the stats confirm the z-direction ports aren't used, so that's good)

May also want to make the 3D flexible enough so that if the z-dim is 1 that we don't have ports 4,5 in existence.